plugins/pdf-research/skills/pdf-research/scripts/requirements.txt
# LightRAG PDF Research System Dependencies
lightrag-hku[api]>=1.4.9
pymupdf>=1.24.0
python-dotenv>=1.0.0
